Implement Delete Approval Limits

See merge request cpsc3192017w2/coast-capital/Caffeinated-Coasters/expense-tracker!23

+++ db/init.sql
@@ -32,7 +32,7 @@ CREATE TABLE employee (
 CREATE TABLE employee_cost_centre (
   employee_id INT NOT NULL,
   cost_centre_id INT NOT NULL,
-  approval_limit INT,
+  approval_limit DECIMAL(10,2),
   PRIMARY KEY (employee_id, cost_centre_id),
   FOREIGN KEY (employee_id) REFERENCES employee(id),
   FOREIGN KEY (cost_centre_id) REFERENCES cost_centre(id)
